melhorias nas funções secundárias

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: rida-cos <rida-cos@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/03 23:17:39 by rdcm              #+#    #+#             */
-/*   Updated: 2025/08/04 21:17:51 by rida-cos         ###   ########.fr       */
+/*   Updated: 2025/08/09 14:45:07 by rida-cos         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ int					ft_putstr(char *s, int fd);
 int					ft_putnbr(int n, int fd);
 int					ft_putptr(void *addr);
 unsigned long long	n_digits(unsigned long long n, size_t base);
-int					ft_puthex(unsigned long long nb, char type);
+int					ft_puthex(unsigned int nb, char type);
 unsigned int		ft_putnbr_unsigned(unsigned int n, int fd);
 
 #endif
